Move Wanted action to new namespace
